Allow dangling faces; Remove wide triangles from hex grid.

+++ landlab/graph/quantity/ext/of_patch.pyx
@@ -26,8 +26,8 @@ cdef calc_area_of_patch(long * nodes_at_patch, long n_vertices,
                         double * x_of_node, double * y_of_node):
     cdef int n
     cdef int node
-    cdef float * x_of_vertex = <float *>malloc(n_vertices * sizeof(float))
-    cdef float * y_of_vertex = <float *>malloc(n_vertices * sizeof(float))
+    cdef double * x_of_vertex = <double *>malloc(n_vertices * sizeof(double))
+    cdef double * y_of_vertex = <double *>malloc(n_vertices * sizeof(double))
 
     try:
         for n in range(n_vertices):
@@ -63,8 +63,8 @@ cdef calc_centroid_of_patch(long * nodes_at_patch, long n_vertices,
                             double * x_of_node, double * y_of_node, double * out):
     cdef int n
     cdef int node
-    cdef float * x = <float *>malloc(n_vertices * sizeof(float))
-    cdef float * y = <float *>malloc(n_vertices * sizeof(float))
+    cdef double * x = <double *>malloc(n_vertices * sizeof(double))
+    cdef double * y = <double *>malloc(n_vertices * sizeof(double))
 
     try:
         for n in range(n_vertices):
@@ -80,7 +80,7 @@ cdef calc_centroid_of_patch(long * nodes_at_patch, long n_vertices,
         free(x)
 
 
-cdef calc_centroid_of_polygon(float * x, float * y, long n_vertices,
+cdef calc_centroid_of_polygon(double * x, double * y, long n_vertices,
                               double * out):
     cdef double x_of_centroid = 0.
     cdef double y_of_centroid = 0.
@@ -108,7 +108,7 @@ cdef calc_centroid_of_polygon(float * x, float * y, long n_vertices,
     out[1] = y_of_centroid
 
 
-cdef calc_area_of_polygon(float * x, float * y, long n_vertices):
+cdef calc_area_of_polygon(double * x, double * y, long n_vertices):
     cdef double area = 0.
     cdef int n
